Create Dog.java
Create a class to represent a Dog, instantiate it and call the methods in it. [1] Create a class called Dog with an overloaded bark( ) method. This method should be overloaded based on various primitive data types, and print different types of barking, howling, etc., depending on which overloaded version is called. Write a main( ) that calls all the different versions.